+++ NEWS.md
@@ -1,3 +1,11 @@
+# 0.0.3 (02-20-2025)
+## Added
+- New functions for cell fate analysis: `PredictPerturbationTime`, `PredictAttractors`, `PredictFates`, and `PredictCommitment`.
+- `VectorFieldCoherence` function to calculate cell-level local coherence scores for the vector field.
+
+## Changes
+- None
+
 # 0.0.2 (09-16-2025)
 ## Added
 - `CustomPerturbation` function.